chore: Refactor CMD function and add Exec function
Refactor the CMD function in command.go to mark it as deprecated and introduce a new Exec function as a replacement. The CMD function now calls the Exec function internally. This change improves code readability and maintainability.

Also, add new options to customize the behavior of the Exec function. The options include setting the working directory, using the parent directory of the executable as the working directory, and disabling console colors in the logger.

This commit addresses the following changes:
- Deprecated the CMD function and introduced the Exec function
- Added options to customize the behavior of the Exec function

Signed-off-by: 117503445 <t117503445@gmail.com>